+++ WebExample/tests/.eslintrc.js
@@ -1,5 +1,6 @@
 module.exports = {
   'rules': {
-    '@lwc/lwc/no-async-await': 'off'
+    '@lwc/lwc/no-async-await': 'off',
+    'rulesdir/prefer-import-module-contents': 'off',
   }
 }
