+++ WebExample/tests/input.spec.ts
@@ -1,13 +1,6 @@
 import {test, expect} from '@playwright/test';
-import type {Page} from '@playwright/test';
 import * as TEST_CONST from '../../testConstants';
-
-const setupInput = async (page: Page, mode: 'clear' | 'reset') => {
-  const inputLocator = await page.locator(`div#MarkdownInput_Example`);
-  await page.click(`[data-testid="${mode}"]`);
-
-  return inputLocator;
-};
+import {checkCursorPosition, setupInput} from './utils';
 
 test.beforeEach(async ({page}) => {
   await page.goto(TEST_CONST.LOCAL_URL, {waitUntil: 'load'});
@@ -31,15 +24,7 @@ test.describe('standard input behaviour', () => {
 
     expect(await inputLocator.innerText()).toBe(EXAMPLE_LONG_CONTENT);
 
-    const cursorPosition = await page.evaluate(() => {
-      const editableDiv = document.querySelector('div[contenteditable="true"]') as HTMLElement;
-      const range = window.getSelection()?.getRangeAt(0);
-      if (!range || !editableDiv) return null;
-      const preCaretRange = range.cloneRange();
-      preCaretRange.selectNodeContents(editableDiv);
-      preCaretRange.setEnd(range.endContainer, range.endOffset);
-      return preCaretRange.toString().length;
-    });
+    const cursorPosition = await page.evaluate(checkCursorPosition);
 
     expect(cursorPosition).toBe(EXAMPLE_LONG_CONTENT.length);
   });
